Created test_marginals.py and moved marginal tests into the file
